Added status effect system
It's a component that you must add to any actor.
Currently does damage on tick if its damage value is set to >0
RemainingTickCount should be set to the number of times you want it to activate over its lifespan.
TickDuration should be set to the time between ticks.
Additional effects such as freezing can be added to Effect Begin then removed at Effect End.